+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 
 - Faces `pgmacs-table-header` and `pgmacs-table-data` are used to display the header and the rows of
   database tables.
+  
+- Variable `pgmacs-row-colors` specifies the colors used for alternating rows in a database table.